Refactor files to adhere to the "single file single class" principle
