+++ src/Contact/ContactFormSubmissionMode.cs
@@ -1,0 +1,10 @@
+namespace Skarp.HubSpotClient.Contact
+{
+    public enum ContactFormSubmissionMode
+    {
+        All,
+        None,
+        Newest,
+        Oldest
+    }
+}
